backend/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
yt-dlp==2025.10.22
flask==3.1.0
flask-cors==6.0.1
pillow==12.0.0
